pyproject.toml
[project]
name = "langgraph-mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.0.3",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.8",
    "mcp>=1.3.0",
    "python-dotenv>=1.0.1",
]
